# docs/swagger.yaml
definitions:
  binance.TickerData:
    properties:
      askPrice:
        type: string
      askQty:
        type: string
      bidPrice:
        type: string
      bidQty:
        type: string
      closeTime:
        type: number
      count:
        type: integer
      firstId:
        type: number
      highPrice:
        type: string
      lastId:
        type: number
      lastPrice:
        type: string
      lastQty:
        type: string
      lowPrice:
        type: string
      openPrice:
        type: string
      openTime:
        type: number
      prevClosePrice:
        type: string
      priceChange:
        type: string
      priceChangePercent:
        type: string
      priceChangePercentFloat:
        type: number
      quoteVolume:
        type: string
      symbol:
        type: string
      volume:
        type: string
      weightedAvgPrice:
        type: string
    type: object
  handler.PairListResponse:
    properties:
      pairs:
        items:
          type: string
        type: array
      refresh_period:
        type: integer
    type: object
  handler.Ticker:
    properties:
      askPrice:
        type: string
      askQty:
        type: string
      bidPrice:
        type: string
      bidQty:
        type: string
      closeTime:
        type: number
      count:
        type: integer
      firstId:
        type: number
      highPrice:
        type: string
      lastId:
        type: number
      lastPrice:
        type: string
      lastQty:
        type: string
      lowPrice:
        type: string
      openPrice:
        type: string
      openTime:
        type: number
      prevClosePrice:
        type: string
      priceChange:
        type: string
      priceChangePercent:
        type: string
      priceChangePercentFloat:
        type: number
      quoteVolume:
        type: string
      symbol:
        type: string
      volume:
        type: string
      weightedAvgPrice:
        type: string
    type: object
info:
  contact: {}
paths:
  /binance/ticker/24hr:
    get:
      description: Retrieve 24-hour ticker data for all trading pairs.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/binance.TickerData'
              type: array
            type: array
      summary: Get 24-hour ticker data
      tags:
      - Binance
  /binance/ticker/24hr/{pair}:
    get:
      description: Retrieve ticker data for a specific trading pair.
      parameters:
      - description: Trading pair symbol (e.g., BTCUSDT)
        in: path
        name: pair
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Ticker'
      summary: Get ticker data for a specific trading pair
      tags:
      - Binance
  /binance/ticker/24hr/gainers:
    get:
      description: Retrieve the top gainers with a specified limit, filtered by ending.
      parameters:
      - description: Limit the number of results
        in: query
        name: limit
        type: integer
      - description: Filter results by ending
        in: query
        name: endingFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/binance.TickerData'
              type: array
            type: array
      summary: Get the top gainers with a specified limit, filtered by ending.
      tags:
      - Binance
  /binance/ticker/24hr/gainers/pairs:
    get:
      description: Retrieve the top gainers with a specified limit, filtered by ending
        and exclusion.
      parameters:
      - description: Limit the number of results
        in: query
        name: limit
        type: integer
      - description: Filter results by ending
        in: query
        name: endingFilter
        type: string
      - description: Exclude results with specific ending
        in: query
        name: exclude
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.PairListResponse'
            type: array
      summary: Get the top gainers with a specified limit, filtered by ending and
        exclusion.
      tags:
      - Binance
  /bybit/ticker/24hr:
    get:
      description: This function fetches the 24-hour ticker data for all trading pairs
        in a given market (Spot, Linear, Option, or Inverse).
      parameters:
      - default: spot
        description: Market type (spot, linear, option, inverse)
        enum:
        - spot
        - linear
        - option
        - inverse
        in: query
        name: market
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of TickerData representing 24-hour ticker information
            for each trading pair
          schema:
            items:
              items:
                $ref: '#/definitions/binance.TickerData'
              type: array
            type: array
        "400":
          description: Invalid market type provided
        "500":
          description: Internal Server Error
      summary: Retrieve 24-hour ticker data for the specified market type in Bybit.
      tags:
      - Bybit
  /bybit/ticker/24hr/{pair}:
    get:
      description: This function fetches the latest ticker information for a specified
        trading pair in a chosen market.
      parameters:
      - description: Trading pair symbol (e.g., BTCUSDT)
        in: path
        name: pair
        required: true
        type: string
      - default: spot
        description: Market type (spot, linear, option, inverse); default is 'spot'
        enum:
        - spot
        - linear
        - option
        - inverse
        in: query
        name: market
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: TickerData object containing detailed ticker information for
            the specified pair
          schema:
            items:
              $ref: '#/definitions/binance.TickerData'
            type: array
        "400":
          description: Invalid market type or trading pair symbol
        "500":
          description: Internal Server Error
      summary: Retrieve ticker data for a specific trading pair.
      tags:
      - Bybit
  /bybit/ticker/24hr/gainers:
    get:
      description: This function fetches trading pairs that have shown significant
        price gains over the last 24 hours in a specified market.
      parameters:
      - default: 500
        description: Limit the number of results; default is 500
        in: query
        name: limit
        type: integer
      - description: Filter results by a specific ending symbol
        in: query
        name: endingFilter
        type: string
      - default: spot
        description: Market type (spot, linear, option, inverse); default is 'spot'
        enum:
        - spot
        - linear
        - option
        - inverse
        in: query
        name: market
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of TickerData representing top gainers
          schema:
            items:
              items:
                $ref: '#/definitions/binance.TickerData'
              type: array
            type: array
        "400":
          description: Invalid market type or query parameters
        "500":
          description: Internal Server Error
      summary: Retrieve top gainers in a specified market with an optional limit and
        ending filter.
      tags:
      - Bybit
  /bybit/ticker/24hr/gainers/pairs:
    get:
      description: This function fetches trading pairs that have shown significant
        price gains over the last 24 hours.
      parameters:
      - default: 100
        description: Limit the number of results; default is 100
        in: query
        name: limit
        type: integer
      - default: USDT
        description: Filter results by a specific ending symbol; default is 'USDT'
        in: query
        name: endingFilter
        type: string
      - default: BNB
        description: Exclude results with a specific ending symbol; default is 'BNB'
        in: query
        name: exclude
        type: string
      - default: spot
        description: Market type (spot, linear, option, inverse); default is 'spot'
        enum:
        - spot
        - linear
        - option
        - inverse
        in: query
        name: market
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of PairListResponse representing top gainers
          schema:
            items:
              $ref: '#/definitions/handler.PairListResponse'
            type: array
        "400":
          description: Invalid market type or query parameters
        "500":
          description: Internal Server Error
      summary: Retrieve top gainers in a specified market with filtering options.
      tags:
      - Bybit
swagger: "2.0"